comment ajax validated

+++ src/main/java/com/boot/app/board/web/comment/CommentApiController.java
@@ -5,11 +5,15 @@ import com.boot.app.board.domain.comment.CommentService;
 import com.boot.app.board.domain.member.Member;
 import com.boot.app.board.web.comment.dto.CommentBasicDto;
 import com.boot.app.board.web.comment.dto.CommentSaveDto;
+import com.boot.app.board.web.comment.dto.CommentUpdateDto;
 import com.boot.app.board.web.login.SessionConst;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
@@ -28,7 +32,7 @@ public class CommentApiController {
 
 
     @GetMapping("/{postId}")
-    public ResponseEntity<Map<String,Object>> CommentList(@PathVariable Long postId,@SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member member) {
+    public ResponseEntity<Map<String,Object>> commentList(@PathVariable Long postId,@SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member member) {
         List<Comment> comments = commentService.commentList(postId);
 
    /*     List<CommentListDto> collect = comments.stream()
@@ -61,7 +65,11 @@ public class CommentApiController {
 
 
     @GetMapping("/{postId}/addModalInfo")
-    public ResponseEntity<Map<String,Object>> CommentModalInfo(@PathVariable Long postId, @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member member){
+    public ResponseEntity<Map<String,Object>> commentAddInfo(@PathVariable Long postId, @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member member){
+
+        if(member==null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
 
         Map<String,Object> result = new HashMap<>();
         result.put("postId",postId);
@@ -70,9 +78,57 @@ public class CommentApiController {
         return new ResponseEntity<>(result,HttpStatus.OK);
     }
 
+    @GetMapping("/{commentId}/updateModalInfo")
+    public ResponseEntity<Map<String,String>> commentUpdateModal(@PathVariable Long commentId){
+
+        Comment oneComment = commentService.findOneComment(commentId);
+
+        Map<String,String> result = new HashMap<>();
+
+        result.put("comment",oneComment.getComment());
+        result.put("memberEmail",oneComment.getMemberEmail());
+
+
+
+        return new ResponseEntity<>(result,HttpStatus.OK);
+    }
+
+    @PutMapping("/{commentId}")
+    public ResponseEntity updateComment(@PathVariable Long commentId,
+                                        @RequestBody @Validated CommentUpdateDto dto, BindingResult bindingResult,
+                                        @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member member){
+
+
+        if(member==null){
+            return new ResponseEntity<>("올바르지 않은 접근", HttpStatus.BAD_REQUEST);
+        }
+
+
+        if(bindingResult.hasErrors()){
+            log.info("errors ={}", bindingResult);
+            return new ResponseEntity<>(bindingResult.getAllErrors() , HttpStatus.BAD_REQUEST);
+
+        }
+
+
+
+        Comment comment = Comment.builder().commentId(commentId).comment(dto.getComment()).build();
+
+
+        commentService.update(comment);
+
+
+        return new ResponseEntity(HttpStatus.OK);
+    }
 
     @PostMapping//ResponseEntity로 selectkey값 받아와서 넘겨줌
-    public ResponseEntity<CommentBasicDto> addComment(@RequestBody CommentSaveDto dto){
+    public ResponseEntity<CommentBasicDto> addComment(@RequestBody CommentSaveDto dto,
+                                                      @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member member){
+
+
+        if(member==null){
+            return  new ResponseEntity<>( HttpStatus.BAD_REQUEST);
+        }
 
         Comment mainComment = Comment.builder()
                 .comment(dto.getComment())
@@ -83,7 +139,7 @@ public class CommentApiController {
 
         Long saveCommentId = commentService.insertMainComment(mainComment);
 
-        Comment findMain = commentService.findMainComment(saveCommentId);
+        Comment findMain = commentService.findOneComment(saveCommentId);
 
         CommentBasicDto result = new CommentBasicDto(findMain.getCommentId(), findMain.getComment(), findMain.getMemberEmail(), findMain.getParent(), findMain.getRegDate());
 
@@ -96,7 +152,11 @@ public class CommentApiController {
 
 
     @DeleteMapping("/{commentId}")
-    public ResponseEntity removeComment(@PathVariable Long commentId){
+    public ResponseEntity removeComment(@PathVariable Long commentId , @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member member){
+
+        if(member==null){
+            return new ResponseEntity(HttpStatus.BAD_REQUEST);
+        }
 
         commentService.remove(commentId);
